% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_to_df.R
\name{board_to_df}
\alias{board_to_df}
\title{Turn a board matrix into a dataframe}
\usage{
board_to_df(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A dataframe
}
\description{
Turn a board matrix into a dataframe
}
