% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brians_brain.R
\name{brians_brain_game}
\alias{brians_brain_game}
\title{Run Brian's brain for a specified number of steps}
\usage{
brians_brain_game(initial_state, steps)
}
\arguments{
\item{initial_state}{The initial state.}

\item{steps}{How many steps to run the model for.}
}
\value{
A dataframe with the simulation results
}
\description{
Run Brian's brain for a specified number of steps
}
