% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brians_brain.R
\name{brians_brain_step}
\alias{brians_brain_step}
\title{One step in Brian's brain}
\usage{
brians_brain_step(m)
}
\arguments{
\item{m}{The current state.}
}
\value{
A matrix representing the subsequent state
}
\description{
One step in Brian's brain
}
