% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langtons_ant.R
\name{create_ant_grid}
\alias{create_ant_grid}
\title{Create a grid for Langton's Ant}
\usage{
create_ant_grid(antpos, axis_size, antorient = "north", state_dist = c(1, 0))
}
\arguments{
\item{antpos}{A vector giving the coordinates of the ant on the grid (row,
column).}

\item{axis_size}{The size of the grid dimensions.}

\item{antorient}{One of "north", "east", "south", and "west". Defaults to
"north" if orientation is invalid.}

\item{state_dist}{Distribution of states as given to \code{create_grid}.}
}
\value{
A \code{matrix} suitable as initial state for Langton's Ant.
}
\description{
Create a grid for Langton's Ant
}
\examples{
initial_state <- create_ant_grid(c(15, 15), "north", 30)
plot_state(initial_state, 3, theme_langtons_ant())
}
