% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_board.R
\name{create_board}
\alias{create_board}
\title{Create a state for a cellular automaton}
\usage{
create_board(n_states, state_dist, dims)
}
\arguments{
\item{n_states}{The number of distinct cell states to use for the grid.}

\item{state_dist}{The distribution of cell states.}

\item{dims}{Vector containing board dimensions (rows, columns).}
}
\value{
A \code{matrix} representing an automaton state.
}
\description{
This function returns a \code{matrix} with \code{n_states} distinct cell states.
The frequency distribution of cell states can be supplied using the
parameter \code{state_dist}.
\code{dims} controls the size of the grid.
}
\examples{
create_grid(5, rep(0.2, 5), c(20, 20))
}
