% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fire.R
\name{create_forest}
\alias{create_forest}
\title{Create a forest to burn}
\usage{
create_forest(axis_size, tree_density)
}
\arguments{
\item{axis_size}{Axis size of the forest.}

\item{tree_density}{How many trees to put on the board relatively.}
}
\value{
A matrix that represents a forest.
}
\description{
Create a forest to burn
}
