% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fire.R
\name{create_forest_board}
\alias{create_forest_board}
\title{Convenience function to create an initial state for the forest fire model}
\usage{
create_forest_board(dims, tree_density)
}
\arguments{
\item{dims}{Vector containing board dimensions (rows, columns).}

\item{tree_density}{The density of the forest given as a probability for the
R base \code{sample} function.}
}
\value{
A matrix that represents a forest on fire.
}
\description{
The forest fire model is often initialized as a grid where the left
most column is "on fire".
This function returns an initial state of this kind.
The forest size and the tree density can be provided as well.
}
\examples{
forest <- create_forest_board(c(20, 20), 0.6)
plot_state(forest, 5, theme_forest_fire())
}
