% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a state grid}
\usage{
create_grid(n_states, state_dist, axis_size)
}
\arguments{
\item{n_states}{The number of distinct cell states to use for the grid.}

\item{state_dist}{The distributaion of cell states.}

\item{axis_size}{The grid dimension in each direction.}
}
\value{
A \code{matrix} representing an automaton state.
}
\description{
This function returns a \code{matrix} with \code{n_states} distinct cell states.
The frequency distribution of cell states can be supplied using the
parameter \code{state_dist}.
Finally, \code{axis_size} controls the size of the grid.
}
\examples{
create_grid(5, rep(0.2, 5), 20)
}
