% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fire.R
\name{forest_fire_game}
\alias{forest_fire_game}
\title{Run the forest fire for a specified number of steps}
\usage{
forest_fire_game(initial_state, steps)
}
\arguments{
\item{initial_state}{The initial forest.}

\item{steps}{How many steps to run the model for.}
}
\value{
A dataframe with the tabular representation of the forest fire
}
\description{
Run the forest fire for a specified number of steps
}
