% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moore_neighborhood.R
\name{get_moore_neighborhood}
\alias{get_moore_neighborhood}
\title{Get the Moore neighborhood of a specified cell}
\usage{
get_moore_neighborhood(i, j, axis_size, periodic = TRUE)
}
\arguments{
\item{i}{Row index}

\item{j}{Column index}

\item{axis_size}{Where the board wraps.}

\item{periodic}{Whether to employ periodic boundary conditions or not.}
}
\value{
A list of coordinates of Moore neighborhood cells.
}
\description{
Get the Moore neighborhood of a specified cell
}
