% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_plot}
\alias{grid_plot}
\title{Plot a grid ABM state}
\usage{
grid_plot(grid_df, marker_size, color_scheme)
}
\arguments{
\item{grid_df}{A dataframe representing the current state.}

\item{marker_size}{How big the markers on the plot should be.}

\item{color_scheme}{What colors to use for each cell state.}
}
\value{
A ggplot2 object of the current state.
}
\description{
Plot a grid ABM state
}
