% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langtons_ant.R
\name{langtons_ant_step}
\alias{langtons_ant_step}
\title{One step in Langton's Ant}
\usage{
langtons_ant_step(m)
}
\arguments{
\item{m}{A \code{matrix} representing the current state.}
}
\value{
A \code{matrix} representing the next state according to Langton's Ant.
}
\description{
Langton's Ant is a cellular automaton where an "ant" moves across the grid
leaving behind black or white cells according to a specific rule.
For more information, check out
\url{https://en.wikipedia.org/wiki/Langton\%27s_ant}.
}
