% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life_game}
\alias{life_game}
\title{Run Conway's life for a specified number of steps}
\usage{
life_game(initial_state, steps)
}
\arguments{
\item{initial_state}{The initial state.}

\item{steps}{How many steps to run the model for.}
}
\value{
A dataframe with the simulation results
}
\description{
Run Conway's life for a specified number of steps
}
