% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life_step}
\alias{life_step}
\title{One step in the Conway's Life automaton}
\usage{
life_step(board)
}
\arguments{
\item{board}{The current state.}
}
\value{
A matrix representing the subsequent state
}
\description{
Conway's Life is perhaps the most famous cellular automaton.
This function evolves a given state by one step applying the life update
rule.
The input \code{matrix} is expected to have the states 0 and 1.
}
\details{
For more information, check out
\url{https://en.wikipedia.org/wiki/Conway\%27s_Game_of_Life}.
}
\examples{
initial_state <- create_board(2, c(0.7, 0.3), c(20, 20))

# Visualize initial state
plot_state(initial_state, 5, theme_life())

next_state <- life_step(initial_state)

# Visualize subsequent state
plot_state(next_state, 5, theme_life())
}
