% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot a grid ABM state}
\usage{
plot_state(state, marker_size, color_scheme)
}
\arguments{
\item{state}{A dataframe or a matrix representing the current state.}

\item{marker_size}{How big the markers on the plot should be.}

\item{color_scheme}{What colors to use for each cell state.}
}
\value{
A ggplot2 object of the current state.
}
\description{
Plot a grid ABM state
}
