% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_automaton.R
\name{run_automaton}
\alias{run_automaton}
\title{Run an automaton with a given stepping function}
\usage{
run_automaton(initial_state, steps, stepfunc, ...)
}
\arguments{
\item{initial_state}{The initial state of the automaton.}

\item{steps}{The number of steps to run the automaton.}

\item{stepfunc}{The stepping function.}

\item{...}{Additional parameters for the stepping function.}
}
\value{
A dataframe with the results of the simulation
}
\description{
Run an automaton with a given stepping function
}
