% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling.R
\name{schelling_game}
\alias{schelling_game}
\title{Run the Schelling model for a specified number of steps}
\usage{
schelling_game(initial_state, tolerance, steps)
}
\arguments{
\item{initial_state}{The initial state of the board.}

\item{tolerance}{The tolerance level.}

\item{steps}{How many steps to run the model.}
}
\value{
A dataframe with the simulation results.
}
\description{
Run the Schelling model for a specified number of steps
}
