% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling.R
\name{schelling_step}
\alias{schelling_step}
\title{Step the Schelling model}
\usage{
schelling_step(board, tolerance)
}
\arguments{
\item{board}{The current state.}

\item{tolerance}{The tolerance level.}
}
\value{
The updated state
}
\description{
Step the Schelling model
}
\examples{
initial_state <- create_board(3, c(0.1, 0.45, 0.45), c(20, 20))

# Visualize initial state
plot_state(initial_state, 5, theme_schelling())

next_state <- schelling_step(initial_state, tolerance = 3)

# Visualize subsequent state
plot_state(next_state, 5, theme_schelling())
}
