% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_brians_brain_dark}
\alias{theme_brians_brain_dark}
\title{Dark theme for the Brian's Brain automaton}
\usage{
theme_brians_brain_dark()
}
\value{
A vector containing the color scheme.
}
\description{
Dark theme for the Brian's Brain automaton
}
