% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_brians_brain_light}
\alias{theme_brians_brain_light}
\title{Light theme for the Brian's Brain automaton}
\usage{
theme_brians_brain_light()
}
\value{
A vector containing the color scheme
}
\description{
Light theme for the Brian's Brain automaton
}
