% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Default theme if no other theme was supplied}
\usage{
theme_default()
}
\value{
A vector containing the color scheme.
}
\description{
Default theme if no other theme was supplied
}
\examples{
theme_default()
}
