% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_forest_fire}
\alias{theme_forest_fire}
\title{Theme for the Forest Fire automaton}
\usage{
theme_forest_fire()
}
\value{
A vector containing the color scheme.
}
\description{
Theme for the Forest Fire automaton
}
\examples{
theme_forest_fire()
}
