% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_langtons_ant}
\alias{theme_langtons_ant}
\title{Theme for Langton's Ant}
\usage{
theme_langtons_ant()
}
\value{
A vector containing the color scheme.
}
\description{
Theme for Langton's Ant
}
\examples{
theme_langtons_ant()
}
