% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_schelling}
\alias{theme_schelling}
\title{Theme for the Schelling automaton}
\usage{
theme_schelling()
}
\value{
A vector containing the color scheme
}
\description{
Theme for the Schelling automaton
}
