% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_model_run.R
\name{animate_model_run}
\alias{animate_model_run}
\title{Animate a model run}
\usage{
animate_model_run(df, marker_size, color_scheme)
}
\arguments{
\item{df}{The simulation data as returned by the \code{run_automaton} function.}

\item{marker_size}{Size of the cells on the grid
(needs adjustment depending on grid size).}

\item{color_scheme}{Which colors to use for which cell state. As the cell
states are defined by a sequence of integers starting at 0, you can simply
specify the colors by providing a vector with each entry \code{i} representing
the color for state \code{i}. To specify colors for specific states, provide a
named vector, where the names are the cell states in backticks (e.g.,
c(\code{0} = "transparent", \code{1} = "black")).}
}
\value{
A gganimate animation.
}
\description{
\code{animate_model_run} takes the results of an automaton run and animates it
leveraging the \code{gganimate} library.
You can specify the marker size on the grid and customize the color scheme.
}
\examples{
initial_state <- create_board(4, rep(0.25, 4), c(20, 20))

results <- run_automaton(initial_state, 100, schelling_step, tolerance = 3)

animate_model_run(
  results,
  5,
  c(`0` = "transparent", `1` = "gold2", `2` = "royalblue", `3` = "coral")
)
}
