% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_model_run.R
\name{animate_model_run}
\alias{animate_model_run}
\title{Animate a model run}
\usage{
animate_model_run(df, marker_size, color_scheme)
}
\arguments{
\item{df}{The simulation data.}

\item{marker_size}{Size of the markers.}

\item{color_scheme}{Which colors to use for which cell state.}
}
\value{
A gganimate animation
}
\description{
Animate a model run
}
