% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fire.R
\name{forest_fire_step}
\alias{forest_fire_step}
\title{Run the forest model one step}
\usage{
forest_fire_step(forest)
}
\arguments{
\item{forest}{A matrix representing a forest}
}
\value{
The updated forest
}
\description{
Run the forest model one step
}
