% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling.R
\name{get_happiness_mat}
\alias{get_happiness_mat}
\title{Compute the happiness of each agent in a Schelling run}
\usage{
get_happiness_mat(board, tolerance)
}
\arguments{
\item{board}{The current state.}

\item{tolerance}{The tolerance level.}
}
\value{
A matrix with corresponding dimensions indicating agent happiness
}
\description{
Compute the happiness of each agent in a Schelling run
}
\examples{
state <- create_board(3, c(0.1, 0.45, 0.45), c(20, 20))
plot_state(state, 5, theme_schelling())
get_happiness_mat(state, 3)
}
