% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moore_neighborhood.R
\name{get_moore_neighborhood}
\alias{get_moore_neighborhood}
\title{Get the Moore neighborhood of a specified cell}
\usage{
get_moore_neighborhood(i, j, dims, periodic = TRUE)
}
\arguments{
\item{i}{Row index.}

\item{j}{Column index.}

\item{periodic}{Whether to employ periodic boundary conditions or not.}

\item{axis_size}{Grid dimension in each direction.}
}
\value{
A list of coordinates of Moore neighborhood cells.
}
\description{
Get the Moore neighborhood of a specified cell
}
\examples{
get_moore_neighborhood(5, 5, 20)
}
