% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_von_neumann_neighborhood.R
\name{get_von_neumann_neighborhood}
\alias{get_von_neumann_neighborhood}
\title{Get Von Neumann neighborhood of a specified cell}
\usage{
get_von_neumann_neighborhood(i, j, axis_size, periodic = TRUE)
}
\arguments{
\item{i}{Row index}

\item{j}{Column index}

\item{axis_size}{Where the board wraps.}

\item{periodic}{Whether to employ periodic boundary conditions or not.}
}
\value{
A list of coordinates of Von Neumann neighborhood cells.
}
\description{
Get Von Neumann neighborhood of a specified cell
}
