% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langtons_ant.R
\name{langtons_ant_step}
\alias{langtons_ant_step}
\title{One step in the Langton's Ant automaton}
\usage{
langtons_ant_step(board)
}
\arguments{
\item{board}{A \code{matrix} representing the current state.}
}
\value{
A \code{matrix} representing the next state according to Langton's Ant.
}
\description{
Langton's Ant is a cellular automaton where an "ant" moves across the grid
leaving behind black or white cells according to a specific rule.
For more information, check out
\url{https://en.wikipedia.org/wiki/Langton\%27s_ant}.
}
\examples{
initial_state <- create_ant_board(c(10, 10), c(20, 20), "north")

# Visualize initial state
plot_state(initial_state, 5, theme_langtons_ant())

next_state <- langtons_ant_step(initial_state)

# Visualize next state
plot_state(next_state, 5, theme_langtons_ant())
}
