% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life_step}
\alias{life_step}
\title{One step in Conway's Life}
\usage{
life_step(m)
}
\arguments{
\item{m}{The current state.}
}
\value{
A matrix representing the subsequent state
}
\description{
Conway's Life is perhaps the most famous cellular automaton.
This function evolves a given state by one step applying the life update
rule.
The input \code{matrix} is expected to have the states 0 and 1.
}
\details{
For more information, check out
\url{https://en.wikipedia.org/wiki/Conway\%27s_Game_of_Life}.
}
\examples{
initial_state <- matrix(
  sample(c(0, 1), prob = c(0.7, 0.3), replace = TRUE, 400),
  nrow = 20, ncol = 20
)

# Visualize initial state
plot_state(initial_state, 5, theme_life())

next_state <- life_step(initial_state)

# Visualize subsequent state
plot_state(next_state, 5, theme_life())
}
