% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life_step}
\alias{life_step}
\title{One step in Conway's life}
\usage{
life_step(m)
}
\arguments{
\item{m}{The current state.}
}
\value{
A matrix representing the subsequent state
}
\description{
One step in Conway's life
}
