% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot a cellular automaton state}
\usage{
plot_state(state, marker_size = 5, color_scheme = theme_default())
}
\arguments{
\item{state}{A \code{data.frame} or a \code{matrix} representing the current state.}

\item{marker_size}{How big the markers on the plot should be.}

\item{color_scheme}{What colors to use for each cell state.}
}
\value{
A \code{ggplot2} object of the current state.
}
\description{
All visualizations in the \code{gridabm} library are \code{ggplot2} objects and can be
customized accordingly.
This function either takes a state \code{matrix} or a \code{data.frame} as returned by
\code{board_to_df} and returns a visualization of the state as a \code{ggplot2} object.
Marker size and color scheme can be customized as well.
The default color palette used is based on the \code{Dark2} palette from
\code{RColorBrewer}.
}
\examples{
state <- create_board(4, c(0.1, 0.3, 0.3, 0.3), c(20, 20))
plot_state(state)
}
