% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling.R
\name{random_free_pos}
\alias{random_free_pos}
\title{Find a random free position in the Schelling game}
\usage{
random_free_pos(board)
}
\arguments{
\item{board}{The current state.}
}
\value{
Coordinates of a random free position
}
\description{
Find a random free position in the Schelling game
}
