% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling.R
\name{schelling_step}
\alias{schelling_step}
\title{Step the Schelling model}
\usage{
schelling_step(board, tolerance)
}
\arguments{
\item{board}{The current state.}

\item{tolerance}{The tolerance level.}
}
\value{
The updated state
}
\description{
Step the Schelling model
}
